requirements.txt
# Core dependencies
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16

# LLM Integration (multiple options)
openai==1.51.0
anthropic==0.34.2

# Code Analysis
gitpython==3.1.40
radon==6.0.1

# Text Processing
tiktoken==0.5.2
chardet==5.2.0
httpx==0.27.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3

# Output formatting
jinja2==3.1.2

# Progress tracking
tqdm==4.66.1

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
